pyproject.toml
[tool.poetry]
name = "librarydjango"
version = "0.1.0"
description = ""
authors = ["goaho7 <goaho7@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
djangorestframework = "^3.15.2"
django-environ = "^0.11.2"
psycopg2 = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
